# docs/api/openapi-fitness.yaml
openapi: 3.1.0
info:
  title: Full Tracker Fitness API
  version: 1.0.0
  description: |
    RESTful API for fitness tracking with support for Wear OS, Health Connect, and mobile applications.

    ## Features
    - Exercise library management
    - Workout session tracking
    - Real-time workout updates
    - Batch synchronization for wearables
    - Health metrics integration
    - Routine templates

    ## Authentication
    This API uses OAuth 2.0 for authentication. Include your access token in the Authorization header:
    ```
    Authorization: Bearer YOUR_ACCESS_TOKEN
    ```

  contact:
    name: API Support
    url: https://fulltracker.app/support
    email: api@fulltracker.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.fulltracker.app/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /workouts:
    get:
      tags: [Workouts]
      summary: List workouts
      description: Get a paginated list of your workout sessions
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkoutStatus'
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
        - name: routine_id
          in: query
          schema:
            type: string
        - name: include
          in: query
          description: Comma-separated list of relationships to include
          schema:
            type: string
            example: exercises,routine,metrics
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of workouts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Workout'

    post:
      tags: [Workouts]
      summary: Create workout
      description: Create a new workout session (start a workout)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutCreate'
      responses:
        '201':
          description: Workout created successfully
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: -created_at
      description: Sort field (prefix with - for descending)
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for safe retries
  schemas:
...
    Workout:
      type: object
      required:
        - id
        - user_id
        - title
        - status
        - started_at
      properties:
        id:
          type: string
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
        routine_id:
          type: string
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExercise'
        notes:
          type: string
        location:
          type: object
          properties:
            type:
              type: string
              enum: [gym, home, outdoor]
            name:
              type: string
        metrics:
          $ref: '#/components/schemas/WorkoutMetrics'
        sync_metadata:
          type: object
          properties:
            source:
              type: string
              enum: [web, wear_os, health_connect, ios]
            device_id:
              type: string
            synced_at:
              type: string
              format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkoutCreate:
      type: object
      required:
        - title
        - started_at
      properties:
        title:
          type: string
        description:
          type: string
        routine_id:
          type: string
        started_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        location:
          type: object
          properties:
            type:
              type: string
              enum: [gym, home, outdoor]
            name:
              type: string
...
    WorkoutStatus:
      type: string
      enum:
        - planned
        - in_progress
        - completed
        - cancelled
    WorkoutMetrics:
      type: object
      properties:
        total_volume:
          type: number
          description: Total weight * reps
        total_reps:
          type: integer
        total_sets:
          type: integer
        calories_burned:
          type: number
        average_heart_rate:
          type: number
        max_heart_rate:
          type: number
    WorkoutExercise:
      type: object
      required:
        - id
        - exercise_id
        - exercise_name
        - order
      properties:
        id:
          type: string
        exercise_id:
          type: string
        exercise_name:
          type: string
        order:
          type: integer
        sets:
          type: array
          items:
            $ref: '#/components/schemas/Set'
        notes:
          type: string
        superset_group:
          type: string
...
    Set:
      type: object
      required:
        - id
        - set_number
        - set_type
      properties:
        id:
          type: string
        workout_exercise_id:
          type: string
        set_number:
          type: integer
        set_type:
          $ref: '#/components/schemas/SetType'
        target_reps:
          type: integer
        actual_reps:
          type: integer
        target_weight:
          type: number
        actual_weight:
          type: number
        target_duration_seconds:
          type: integer
        actual_duration_seconds:
          type: integer
        target_distance_meters:
          type: number
        actual_distance_meters:
          type: number
        rpe:
          type: integer
          minimum: 1
          maximum: 10
          description: Rate of Perceived Exertion
        rest_seconds:
          type: integer
        completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
        notes:
          type: string
...
    SetType:
      type: string
      enum:
        - normal
        - warmup
        - drop_set
        - superset
        - rest_pause
        - failure
        - amrap

    # ==================== ROUTINE ====================
...
    PaginatedResponse:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
            has_next:
              type: boolean
            has_prev:
              type: boolean
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        api_version:
          type: string
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        field_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
        documentation_url:
          type: string

    # ==================== WRAPPER RESPONSES ====================
...
    WorkoutResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Workout'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
        meta:
          $ref: '#/components/schemas/Meta'
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...